Lerper'

+++ panoramator/src/main/java/space/itoncek/Snapshot3D.java
@@ -6,7 +6,7 @@ import java.time.LocalDateTime;
 
 import static space.itoncek.Panoramator.*;
 
-public record Snapshot3D(double azi, double alt, int fov) {
+public record Snapshot3D(double azi, double alt, double fov) {
     Snapshot5D convertTo5D(LocalDateTime time) {
         return new Snapshot5D(this.azi, this.alt, this.fov, integerPart(julian(time)), fractionalPart(julian(time)));
     };
